% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ziff.R
\name{get_ziff}
\alias{get_ziff}
\title{Return raw commercial landings data from a ZIFF (Zonal Interchange Format File)}
\usage{
get_ziff(years, nafo)
}
\arguments{
\item{years}{Which year(s) to return.}

\item{nafo}{Which NAFO divistion(s). Case insensitive.}
}
\description{
Return raw commercial landings data from a ZIFF (Zonal Interchange Format File)
}
\examples{
#get_ziff(years=2022:2023,nafo=c("4r","4s","4t","4vn",'4vs',"3pn",'3ps'))
}
